Prints the minimum number of coins

+++ 0x0A-argc_argv/100-change.c
@@ -10,8 +10,8 @@
 
 int main(int argc, char *argv[])
 {
-	long cents;
-	long coins = 0;
+	int cents;
+	int coins = 0;
 
 	if (argc != 2)
 	{
